+++ modules/item/item.js
@@ -1,7 +1,6 @@
 import { HMTABLES, HMCONST } from '../tables/constants.js';
 import { HMChatMgr } from '../mgr/chatmgr.js';
 import { HMDialogMgr } from '../mgr/dialogmgr.js';
-import { HMRollMgr } from '../mgr/rollmgr.js';
 import { HMStates } from '../sys/effects.js';
 
 // Remember: Items may not alter Actors under any circumstances.
@@ -107,47 +106,6 @@ export class HMItem extends Item {
             const card = await chatMgr.getCard({dataset});
             await ChatMessage.create(card);
         });
-    }
-
-    static async createItem(item, _options, userId) {
-        if (game.user.id !== userId) return;
-        const {parent, type} = item;
-
-        if (type === 'wound') {
-            if (!parent.system.bonus.total.trauma) return;
-            const {top} = parent.system.hp;
-            const wound = item.system.hp;
-            if (!top || top >= wound) return;
-
-            const chatmgr = new HMChatMgr();
-            const cardtype = HMCONST.CARD_TYPE.ALERT;
-            const dataset = {context: item, top, wound};
-            if (parent.type === 'beast') dataset.hidden = true;
-            const card = await chatmgr.getCard({cardtype, dataset});
-            await ChatMessage.create(card);
-
-            // Auto-roll for beasts.
-            if (parent.type === 'beast') {
-                dataset.dialog = 'save';
-                dataset.formulaType = 'trauma';
-
-                const rollMgr = new HMRollMgr();
-                const dialogResp = {caller: parent, context: parent};
-                dataset.resp = {caller: parent};
-
-                const roll = await rollMgr.getRoll(dataset, dialogResp);
-                const rollMode = 'gmroll';
-                dialogResp.resp = {rollMode};
-                const topcard = await chatmgr.getCard({dataset, roll, dialogResp});
-                await ChatMessage.create(topcard);
-            }
-        } else if (type === 'race' || type === 'cclass') {
-            const itemList = parent.itemTypes[type];
-            if (!itemList.length) return;
-
-            Object.entries(itemList.slice(0, itemList.length -1))
-                .map((a) => parent.items.get(a[1].id).delete());
-        }
     }
 
     onClick() {